feat: PostDemandeDTO

+++ src/main/java/com/gestionEtudiants/Dao/ServiceRepository.java
@@ -3,5 +3,5 @@ package com.gestionEtudiants.Dao;
 import com.gestionEtudiants.Entites.Service;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceRepository extends JpaRepository<Service, Long> {
+public interface ServiceRepository extends JpaRepository<Service, Integer> {
 }
